fix : entity rename, fix plots
