+++ src/main/java/f1b4/webide_server/dto/DeleteFileReq.java
@@ -1,12 +1,8 @@
 package f1b4.webide_server.dto;
 
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 
-import java.lang.reflect.Member;
-
-@Getter
-@Setter
+@Data
 public class DeleteFileReq {
 
     private Long id;
