+++ src/main/java/f1b4/webide_server/dto/ReadFileReq.java
@@ -1,10 +1,8 @@
 package f1b4.webide_server.dto;
 
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 
-@Getter
-@Setter
+@Data
 public class ReadFileReq {
     private String fileName;
     //private Long memeberId;
