+++ src/main/java/f1b4/webide_server/repository/FileRepository.java
@@ -1,6 +1,6 @@
 package f1b4.webide_server.repository;
 
-import f1b4.webide_server.domain.FileEntity;
+import f1b4.webide_server.entity.FileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
